requirements.txt
flask
nltk
scikit-learn
psycopg2-binary
